+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - All the message structs now have both `Serialize` and `Deserialize`
+- Device types from Stream Deck software version 6.0 (Stream Deck +).
 
 
 ## [0.5.1] - 2022-08-13